Moved getCoursesByTerm to BaseViewModel
